chore(signing): rename s3 and polly signers
