+++ packages/polly-signer/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./getPresignedSynthesizeSpeechUrl";
+export * from "./getSignedUrl";
