+++ packages/s3-request-presigner/package.json
@@ -9,7 +9,7 @@
     "build:types": "tsc -p tsconfig.types.json",
     "build:types:downlevel": "downlevel-dts dist-types dist-types/ts3.4",
     "clean": "rimraf ./dist-* && rimraf *.tsbuildinfo",
-    "test": "jest"
+    "test": "jest --passWithNoTests"
   },
   "main": "./dist-cjs/index.js",
   "module": "./dist-es/index.js",
@@ -23,6 +23,7 @@
     "@aws-sdk/middleware-endpoint": "*",
     "@aws-sdk/middleware-sdk-s3": "*",
     "@aws-sdk/protocol-http": "*",
+    "@aws-sdk/s3-signer": "*",
     "@aws-sdk/signature-v4-multi-region": "*",
     "@aws-sdk/smithy-client": "*",
     "@aws-sdk/types": "*",
