+++ packages/s3-signer/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@aws-sdk/polly-request-presigner",
-  "version": "3.186.0",
+  "name": "@aws-sdk/s3-signer",
+  "version": "3.0.0",
   "scripts": {
     "build": "concurrently 'yarn:build:cjs' 'yarn:build:es' 'yarn:build:types'",
     "build:cjs": "tsc -p tsconfig.cjs.json",
@@ -9,7 +9,7 @@
     "build:types": "tsc -p tsconfig.types.json",
     "build:types:downlevel": "downlevel-dts dist-types dist-types/ts3.4",
     "clean": "rimraf ./dist-* && rimraf *.tsbuildinfo",
-    "test": "jest --passWithNoTests"
+    "test": "jest"
   },
   "main": "./dist-cjs/index.js",
   "module": "./dist-es/index.js",
@@ -20,16 +20,18 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@aws-sdk/client-polly": "*",
-    "@aws-sdk/polly-signer": "*",
+    "@aws-sdk/middleware-endpoint": "*",
+    "@aws-sdk/middleware-sdk-s3": "*",
     "@aws-sdk/protocol-http": "*",
-    "@aws-sdk/signature-v4": "*",
+    "@aws-sdk/signature-v4-multi-region": "*",
+    "@aws-sdk/smithy-client": "*",
     "@aws-sdk/types": "*",
     "@aws-sdk/util-create-request": "*",
     "@aws-sdk/util-format-url": "*",
     "tslib": "^2.3.1"
   },
   "devDependencies": {
+    "@aws-sdk/client-s3": "*",
     "@aws-sdk/hash-node": "*",
     "@tsconfig/recommended": "1.0.1",
     "@types/node": "^12.0.2",
@@ -52,10 +54,10 @@
   "files": [
     "dist-*"
   ],
-  "homepage": "https://github.com/aws/aws-sdk-js-v3/tree/main/packages/polly-request-presigner",
+  "homepage": "https://github.com/aws/aws-sdk-js-v3/tree/main/packages/s3-signer",
   "repository": {
     "type": "git",
     "url": "https://github.com/aws/aws-sdk-js-v3.git",
-    "directory": "packages/polly-request-presigner"
+    "directory": "packages/s3-signer"
   }
 }
