+++ packages/s3-signer/src/constants.ts
