+++ packages/s3-signer/src/presigner.spec.ts
