+++ scripts/benchmark-size/scope.json
@@ -45,9 +45,9 @@
   { "package": "@aws-sdk/abort-controller" },
   { "package": "@aws-sdk/fetch-http-handler" },
   { "package": "@aws-sdk/node-http-handler", "skipBundlerTests": true },
-  { "package": "@aws-sdk/polly-request-presigner" },
+  { "package": "@aws-sdk/polly-signer" },
   { "package": "@aws-sdk/s3-presigned-post" },
-  { "package": "@aws-sdk/s3-request-presigner" },
+  { "package": "@aws-sdk/s3-signer" },
   { "package": "@aws-sdk/signature-v4" },
   { "package": "@aws-sdk/signature-v4-crt", "skipBundlerTests": true },
   { "package": "@aws-sdk/smithy-client" },
